chore: move enums to types
