--- apps/frontend/src/app/5_pages/ProtocolDataPage/components/EcosystemStatistics/EcosystemStatistics.constants.ts
@@ -1,8 +1,0 @@
-export enum EcosystemContracts {
-  ProtocolContracts = 'protocolContracts',
-  LendingContracts = 'lendingContracts',
-  AmmContracts = 'ammContracts',
-  ZeroContracts = 'zeroContracts',
-  MyntAggregatorContracts = 'myntAggregatorContracts',
-  BitocracyStakingContract = 'bitocracyStakingContract',
-}
