+++ apps/frontend/src/app/5_pages/ProtocolDataPage/components/EcosystemStatistics/EcosystemStatistics.types.ts
@@ -7,3 +7,12 @@ export type EcosystemStatisticsProps = {
 };
 
 export type PoolData = Record<string, Decimal>;
+
+export enum EcosystemContracts {
+  ProtocolContracts = 'protocolContracts',
+  LendingContracts = 'lendingContracts',
+  AmmContracts = 'ammContracts',
+  ZeroContracts = 'zeroContracts',
+  MyntAggregatorContracts = 'myntAggregatorContracts',
+  BitocracyStakingContract = 'bitocracyStakingContract',
+}
