+++ apps/frontend/src/app/5_pages/ProtocolDataPage/components/EcosystemStatistics/components/BitocracyStakingContract/BitocracyStakingContract.tsx
@@ -12,7 +12,7 @@ import {
   getConvertedValue,
   getCurrencyPrecision,
 } from '../../../../../PortfolioPage/components/ProtocolSection/ProtocolSection.utils';
-import { EcosystemContracts } from '../../EcosystemStatistics.constants';
+import { EcosystemContracts } from '../../EcosystemStatistics.types';
 import { EcosystemStatisticsProps } from '../../EcosystemStatistics.types';
 
 export const BitocracyStakingContract: FC<EcosystemStatisticsProps> = ({
