+++ apps/frontend/src/app/5_pages/ProtocolDataPage/components/EcosystemStatistics/components/LendingContracts/LendingContracts.tsx
@@ -10,7 +10,7 @@ import {
   getConvertedValue,
   getCurrencyPrecision,
 } from '../../../../../PortfolioPage/components/ProtocolSection/ProtocolSection.utils';
-import { EcosystemContracts } from '../../EcosystemStatistics.constants';
+import { EcosystemContracts } from '../../EcosystemStatistics.types';
 import {
   PoolData,
   EcosystemStatisticsProps,
